+++ source/dotnet/wholesale-api/WebApi/Startup.cs
@@ -29,6 +29,7 @@ using Energinet.DataHub.Wholesale.Common.Infrastructure.Options;
 using Energinet.DataHub.Wholesale.Common.Infrastructure.Security;
 using Energinet.DataHub.Wholesale.Edi.Extensions.DependencyInjection;
 using Energinet.DataHub.Wholesale.WebApi.Extensions.DependencyInjection;
+using Energinet.DataHub.Wholesale.WebApi.Extensions.Options;
 using Microsoft.Extensions.Azure;
 using Microsoft.OpenApi.Models;
 
@@ -63,11 +64,13 @@ public class Startup
             .AddControllers(options => options.Filters.Add<BusinessValidationExceptionFilter>())
             .AddJsonOptions(options => { options.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter()); });
 
-        // Register the Swagger generator, defining 1 or more Swagger documents.
+        // Open API generation
+        // The following article is a good read and shows many of the pieces that we have used: https://medium.com/@mo.esmp/api-versioning-and-swagger-in-asp-net-core-7-0-fe45f67d8419
+        // TODO: EDI does this differently but the following has some similar parts: https://learn.microsoft.com/en-us/aspnet/core/tutorials/getting-started-with-swashbuckle?view=aspnetcore-8.0&tabs=visual-studio#xml-comments
+        services.ConfigureOptions<ConfigureSwaggerOptions>();
         services.AddSwaggerGen(options =>
         {
             options.SupportNonNullableReferenceTypes();
-            options.OperationFilter<BinaryContentFilter>();
 
             // Set the comments path for the Swagger JSON and UI.
             // See: https://learn.microsoft.com/en-us/aspnet/core/tutorials/getting-started-with-swashbuckle?view=aspnetcore-8.0&tabs=visual-studio#xml-comments
@@ -86,22 +89,30 @@ public class Startup
             };
             options.AddSecurityDefinition("Bearer", securitySchema);
 
+            // TODO: EDI does this differently and so does: https://learn.microsoft.com/en-us/aspnet/core/tutorials/getting-started-with-swashbuckle?view=aspnetcore-8.0&tabs=visual-studio#xml-comments
             var securityRequirement = new OpenApiSecurityRequirement { { securitySchema, new[] { "Bearer" } }, };
             options.AddSecurityRequirement(securityRequirement);
+
+            // TODO: Wholesale specific
+            // Support binary content, e.g. for Settlement download
+            options.OperationFilter<BinaryContentFilter>();
         });
 
-        var apiVersioningBuilder = services.AddApiVersioning(options =>
-        {
-            options.DefaultApiVersion = new ApiVersion(3, 0);
-            options.AssumeDefaultVersionWhenUnspecified = true;
-            options.ReportApiVersions = true;
-        });
-        apiVersioningBuilder.AddApiExplorer(options =>
-        {
-            options.GroupNameFormat = "'v'VVV";
-            options.SubstituteApiVersionInUrl = true;
-        });
-        services.ConfigureOptions<ConfigureSwaggerOptions>();
+        // API versioning
+        services
+            .AddApiVersioning(options =>
+            {
+                // If client doesn't specify version, we assume the following default
+                options.DefaultApiVersion = new ApiVersion(3, 0);
+                options.AssumeDefaultVersionWhenUnspecified = true;
+                options.ReportApiVersions = true;
+            })
+            .AddApiExplorer(options =>
+            {
+                // API version format strings: https://github.com/dotnet/aspnet-api-versioning/wiki/Version-Format#custom-api-version-format-strings
+                options.GroupNameFormat = "'v'VVV";
+                options.SubstituteApiVersionInUrl = true;
+            });
 
         // Authentication/authorization
         services
@@ -157,8 +168,8 @@ public class Startup
             {
                 // GroupName is the version (e.g. 'v1') as configured using the AddApiExplorer and the 'GroupNameFormat' property.
                 options.SwaggerEndpoint(
-                    $"/swagger/{description.GroupName}/swagger.json",
-                    description.GroupName.ToUpperInvariant());
+                    url: $"/swagger/{description.GroupName}/swagger.json",
+                    name: description.GroupName.ToUpperInvariant());
             }
         });
